+++ js-packages/@prestojs/viewmodel/src/fields/IPAddressField.ts
@@ -1,9 +1,0 @@
-import CharField from './CharField';
-
-/**
- * @extract-docs
- * @menu-group Fields
- */
-export default class IPAddressField extends CharField {
-    static fieldClassName = 'IPAddressField';
-}