+++ js-packages/@prestojs/viewmodel/src/fields/UUIDField.ts
@@ -1,9 +1,0 @@
-import CharField from './CharField';
-
-/**
- * @extract-docs
- * @menu-group Fields
- */
-export default class UUIDField extends CharField {
-    static fieldClassName = 'UUIDField';
-}